refactor: consolidate event runtime internals and reduce root-level surface
