+++ internal/core/event/api.go
@@ -1,0 +1,23 @@
+package event
+
+import engine "github.com/goplus/spx/v2/pkg/spx/pkg/engine"
+
+type Key = engine.KeyCode
+type Direction = float64
+type BackdropName = string
+
+type IEventSinks interface {
+	OnAnyKey(onKey func(key Key))
+	OnBackdrop__0(onBackdrop func(name BackdropName))
+	OnBackdrop__1(name BackdropName, onBackdrop func())
+	OnClick(onClick func())
+	OnKey__0(key Key, onKey func())
+	OnKey__1(keys []Key, onKey func(Key))
+	OnKey__2(keys []Key, onKey func())
+	OnMsg__0(onMsg func(msg string, data any))
+	OnMsg__1(msg string, onMsg func())
+	OnStart(onStart func())
+	OnSwipe__0(direction Direction, onSwipe func())
+	OnTimer(time float64, onTimer func())
+	Stop(kind StopKind)
+}
