+++ runtime.go
@@ -18,6 +18,7 @@ package spx
 
 import (
 	"github.com/goplus/spbase/mathf"
+	coreevent "github.com/goplus/spx/v2/internal/core/event"
 	coreruntime "github.com/goplus/spx/v2/internal/core/runtime"
 	"github.com/goplus/spx/v2/internal/coroutine"
 	"github.com/goplus/spx/v2/internal/engine"
@@ -40,6 +41,209 @@ func (p *SpriteImpl) engine() *engineManagers {
 
 func (c *componentBase) engine() *engineManagers {
 	return c.sprite.engine()
+}
+
+type threadObj = coroutine.ThreadObj
+type eventSink = coreevent.Sink
+
+type scriptEventBindings struct {
+	*scriptEventRegistry
+	pthis threadObj
+}
+
+type scriptEventRegistry struct {
+	coreevent.Manager
+}
+
+func (p *scriptEventBindings) init(mgr *scriptEventRegistry, this threadObj) {
+	p.scriptEventRegistry = mgr
+	p.pthis = this
+}
+
+func (p *scriptEventBindings) initFrom(src *scriptEventBindings, this threadObj) {
+	p.scriptEventRegistry = src.scriptEventRegistry
+	p.pthis = this
+}
+
+func (p *scriptEventBindings) doDeleteClone() {
+	p.scriptEventRegistry.DeleteOwner(p.pthis)
+}
+
+func (p *scriptEventBindings) doWhenSwipe(direction Direction, target threadObj) {
+	p.scriptEventRegistry.doWhenSwipe(direction, target)
+}
+
+func (p *scriptEventBindings) onAwake(onAwake func()) {
+	pthis := p.pthis
+	p.scriptEventRegistry.AddAwake(coreevent.NewSink(p.pthis, onAwake, coreevent.MatchOwnerOrNil(pthis)))
+}
+
+func nameOf(this any) string {
+	if spr, ok := this.(*SpriteImpl); ok {
+		return spr.name
+	}
+	if _, ok := this.(*Game); ok {
+		return "Game"
+	}
+	engine.Panic("scriptEventBindings: unexpected this object")
+	return ""
+}
+
+func isGame(obj threadObj) bool {
+	_, ok := obj.(*Game)
+	return ok
+}
+
+func isSprite(obj threadObj) bool {
+	_, ok := obj.(*SpriteImpl)
+	return ok
+}
+
+func (p *scriptEventBindings) OnStart(onStart func()) {
+	sink := coreevent.NewSink(p.pthis, onStart)
+	if p.scriptEventRegistry.TryAddStart(sink) {
+		return
+	}
+	if sprite, ok := sink.Owner.(*SpriteImpl); ok && sprite.spriteState.Cloned {
+		return
+	}
+	spxlog.Warn("event: ignoring late OnStart registration for %s", nameOf(sink.Owner))
+}
+
+func (p *scriptEventBindings) OnClick(onClick func()) {
+	pthis := p.pthis
+	p.scriptEventRegistry.AddClick(coreevent.NewSink(pthis, onClick, coreevent.MatchOwner(pthis)))
+}
+
+func (p *scriptEventBindings) OnAnyKey(onKey func(key Key)) {
+	p.scriptEventRegistry.AddKeyPressed(coreevent.NewSink(p.pthis, onKey))
+}
+
+func (p *scriptEventBindings) OnTimer(time float64, call func()) {
+	timer.RegisterTimer(time)
+	p.scriptEventRegistry.AddTimer(coreevent.NewSink(
+		p.pthis,
+		coreevent.TapVoid1(call, coreevent.If1(isDebugEventEnabled, func(float64) {
+			spxlog.Debug("==> onTimer: %s", nameOf(p.pthis))
+		})),
+		coreevent.MatchApproxFloat(time, 0.001),
+	))
+}
+
+func (p *scriptEventBindings) OnKey__0(key Key, onKey func()) {
+	p.scriptEventRegistry.AddKeyPressed(coreevent.NewSink(
+		p.pthis,
+		coreevent.TapVoid1(onKey, coreevent.If1(isDebugEventEnabled, func(Key) {
+			spxlog.Debug("==> onKey: %v, %s", key, nameOf(p.pthis))
+		})),
+		coreevent.MatchValue(key),
+	))
+}
+
+func (p *scriptEventBindings) OnSwipe__0(direction Direction, onSwipe func()) {
+	p.scriptEventRegistry.AddSwipe(coreevent.NewSink(
+		p.pthis,
+		coreevent.TapVoid1(onSwipe, coreevent.If1(isDebugEventEnabled, func(Direction) {
+			spxlog.Debug("==> onSwipe: %v, %s", direction, nameOf(p.pthis))
+		})),
+		coreevent.MatchValue(direction),
+	))
+}
+
+func (p *scriptEventBindings) OnKey__1(keys []Key, onKey func(Key)) {
+	p.scriptEventRegistry.AddKeyPressed(coreevent.NewSink(
+		p.pthis,
+		coreevent.Tap1(onKey, coreevent.If1(isDebugEventEnabled, func(key Key) {
+			spxlog.Debug("==> onKey: %v, %s", keys, nameOf(p.pthis))
+		})),
+		coreevent.MatchAnyOf(keys),
+	))
+}
+
+func (p *scriptEventBindings) OnKey__2(keys []Key, onKey func()) {
+	p.OnKey__1(keys, coreevent.Ignore1[Key](onKey))
+}
+
+func (p *scriptEventBindings) OnMsg__0(onMsg func(msg string, data any)) {
+	p.scriptEventRegistry.AddIReceive(coreevent.NewSink(p.pthis, onMsg))
+}
+
+func (p *scriptEventBindings) OnMsg__1(msg string, onMsg func()) {
+	p.scriptEventRegistry.AddIReceive(coreevent.NewSink(
+		p.pthis,
+		coreevent.TapVoid2(onMsg, coreevent.If2(isDebugEventEnabled, func(msg string, data any) {
+			spxlog.Debug("==> onMsg: %s, %s", msg, nameOf(p.pthis))
+		})),
+		coreevent.MatchValue(msg),
+	))
+}
+
+func (p *scriptEventBindings) OnBackdrop__0(onBackdrop func(name BackdropName)) {
+	p.scriptEventRegistry.AddBackdropChanged(coreevent.NewSink(p.pthis, onBackdrop))
+}
+
+func (p *scriptEventBindings) OnBackdrop__1(name BackdropName, onBackdrop func()) {
+	p.scriptEventRegistry.AddBackdropChanged(coreevent.NewSink(
+		p.pthis,
+		coreevent.TapVoid1(onBackdrop, coreevent.If1(isDebugEventEnabled, func(name BackdropName) {
+			spxlog.Debug("==> onBackdrop: %s, %s", name, nameOf(p.pthis))
+		})),
+		coreevent.MatchValue(name),
+	))
+}
+
+func (p *scriptEventBindings) Stop(kind StopKind) {
+	current := gco.Current()
+	filter, abort := coreevent.ResolveStop(
+		kind,
+		p.pthis,
+		func(obj any) bool { return isSprite(obj) },
+		func(obj any) bool { return isGame(obj) },
+	)
+	if filter != nil {
+		gco.StopIf(func(th coroutine.Thread) bool {
+			return filter(th.Obj, th == current)
+		})
+	}
+	if abort {
+		gco.Abort()
+	}
+}
+
+type eventQueuePolicy = coreevent.QueuePolicy
+
+const defaultEventQueuePolicy = coreevent.DefaultQueuePolicy
+
+type eventQueueSnapshot = coreevent.QueueSnapshot
+
+func parseEventQueuePolicy(policy string) eventQueuePolicy {
+	return coreevent.ParsePolicy(policy)
+}
+
+func (p *Game) initEventQueueState() {
+	p.gameRuntimeState.EventQueuePolicy = defaultEventQueuePolicy
+	p.gameRuntimeState.EventQueueStats.Reset()
+}
+
+func (p *Game) resetEventQueueStats() {
+	p.gameRuntimeState.EventQueueStats.Reset()
+}
+
+func (p *Game) setEventQueuePolicy(policy eventQueuePolicy) {
+	p.gameRuntimeState.EventQueuePolicy = policy
+}
+
+func (p *Game) eventQueueSnapshot() eventQueueSnapshot {
+	queueLen, queueCap := 0, 0
+	if p.events != nil {
+		queueLen = len(p.events)
+		queueCap = cap(p.events)
+	}
+	return coreevent.Snapshot(p.gameRuntimeState.EventQueuePolicy, &p.gameRuntimeState.EventQueueStats, queueLen, queueCap)
+}
+
+func (p *Game) queueEventWithPolicy(ev event) bool {
+	return coreevent.EnqueueWithPolicy(p.events, ev, p.gameRuntimeState.EventQueuePolicy, &p.gameRuntimeState.EventQueueStats, &p.gameRuntimeState.EventQueueMu)
 }
 
 var (
@@ -132,12 +336,12 @@ func (p *Game) handleEvent(ev event) {
 		p.doWhenLeftButtonDown(e)
 	case *eventKeyDown:
 		// Note: key-up callbacks are not part of the current event sink API.
-		p.sinkMgr.doWhenKeyPressed(e.Key)
+		p.scriptEvents.doWhenKeyPressed(e.Key)
 	case *eventStart:
-		p.sinkMgr.doWhenAwake(nil)
-		p.sinkMgr.doWhenStart()
+		p.scriptEvents.doWhenAwake(nil)
+		p.scriptEvents.doWhenStart()
 	case *eventTimer:
-		p.sinkMgr.doWhenTimer(e.Time)
+		p.scriptEvents.doWhenTimer(e.Time)
 	}
 }
 
@@ -147,6 +351,128 @@ func (p *Game) fireEvent(ev event) {
 	}
 	if isDebugInstrEnabled() {
 		spxlog.Warn("Event buffer is full (policy=%s). Drop event: %v", p.gameRuntimeState.EventQueuePolicy, ev)
+	}
+}
+
+func (p *scriptEventRegistry) doWhenStart() {
+	sinks := p.SnapshotStartOnce()
+	if len(sinks) == 0 {
+		return
+	}
+	coreevent.DispatchAsync(sinks, false, nil, eventDispatchHooks(), func(ev *eventSink) {
+		coreevent.If0(isDebugEventEnabled, func() {
+			spxlog.Debug("==> onStart: %s", nameOf(ev.Owner))
+		})()
+		ev.Handler.(func())()
+	})
+}
+
+func (p *scriptEventRegistry) doWhenAwake(this threadObj) {
+	sinks := p.SnapshotAwake()
+	coreevent.DispatchSync(sinks, this, eventDispatchHooks(), func(ev *eventSink) {
+		coreevent.If0(isDebugEventEnabled, func() {
+			spxlog.Debug("==> onAwake: %s", nameOf(ev.Owner))
+		})()
+		ev.Handler.(func())()
+	})
+}
+
+func (p *scriptEventRegistry) doWhenTimer(time float64) {
+	sinks := p.SnapshotTimer()
+	coreevent.DispatchAsync(sinks, false, time, eventDispatchHooks(), func(ev *eventSink) {
+		ev.Handler.(func(float64))(time)
+	})
+}
+
+func (p *scriptEventRegistry) doWhenKeyPressed(key Key) {
+	sinks := p.SnapshotKeyPressed()
+	coreevent.DispatchAsync(sinks, false, key, eventDispatchHooks(), func(ev *eventSink) {
+		ev.Handler.(func(Key))(key)
+	})
+}
+
+func (p *scriptEventRegistry) doWhenSwipe(direction Direction, this threadObj) {
+	sinks := p.SnapshotSwipe()
+	coreevent.DispatchAsync(sinks, false, direction, eventDispatchHooks(), func(ev *eventSink) {
+		if ev.Owner == this {
+			ev.Handler.(func(Direction))(direction)
+		}
+	})
+}
+
+func (p *scriptEventRegistry) doWhenClick(this threadObj) {
+	sinks := p.SnapshotClick()
+	coreevent.DispatchAsync(sinks, false, this, eventDispatchHooks(), func(ev *eventSink) {
+		coreevent.If0(isDebugEventEnabled, func() {
+			spxlog.Debug("==> onClick: %s", nameOf(this))
+		})()
+		ev.Handler.(func())()
+	})
+}
+
+func (p *scriptEventRegistry) doWhenTouchStart(this threadObj, obj *SpriteImpl) {
+	sinks := p.SnapshotTouchStart()
+	coreevent.DispatchAsync(sinks, false, this, eventDispatchHooks(), func(ev *eventSink) {
+		coreevent.If0(isDebugEventEnabled, func() {
+			spxlog.Debug("===> onTouchStart: %s, %s", nameOf(this), obj.name)
+		})()
+		ev.Handler.(func(Sprite))(obj.sprite)
+	})
+}
+
+func (p *scriptEventRegistry) doWhenTouching(this threadObj, obj *SpriteImpl) {
+	sinks := p.SnapshotTouching()
+	coreevent.DispatchAsync(sinks, false, this, eventDispatchHooks(), func(ev *eventSink) {
+		coreevent.If0(isDebugEventEnabled, func() {
+			spxlog.Debug("==> onTouching: %s, %s", nameOf(this), obj.name)
+		})()
+		ev.Handler.(func(Sprite))(obj.sprite)
+	})
+}
+
+func (p *scriptEventRegistry) doWhenTouchEnd(this threadObj, obj *SpriteImpl) {
+	sinks := p.SnapshotTouchEnd()
+	coreevent.DispatchAsync(sinks, false, this, eventDispatchHooks(), func(ev *eventSink) {
+		coreevent.If0(isDebugEventEnabled, func() {
+			spxlog.Debug("===> onTouchEnd: %s, %s", nameOf(this), obj.name)
+		})()
+		ev.Handler.(func(Sprite))(obj.sprite)
+	})
+}
+
+func (p *scriptEventRegistry) doWhenCloned(this threadObj, data any) {
+	sinks := p.SnapshotCloned()
+	coreevent.DispatchAsync(sinks, true, this, eventDispatchHooks(), func(ev *eventSink) {
+		coreevent.If0(isDebugEventEnabled, func() {
+			spxlog.Debug("==> onCloned: %s", nameOf(this))
+		})()
+		ev.Handler.(func(any))(data)
+	})
+}
+
+func (p *scriptEventRegistry) doWhenIReceive(msg string, data any, wait bool) {
+	sinks := p.SnapshotIReceive()
+	coreevent.Dispatch(sinks, wait, msg, eventDispatchHooks(), func(ev *eventSink) {
+		ev.Handler.(func(string, any))(msg, data)
+	})
+}
+
+func (p *scriptEventRegistry) doWhenBackdropChanged(name BackdropName, wait bool) {
+	sinks := p.SnapshotBackdropChanged()
+	coreevent.Dispatch(sinks, wait, name, eventDispatchHooks(), func(ev *eventSink) {
+		ev.Handler.(func(BackdropName))(name)
+	})
+}
+
+func eventDispatchHooks() coreevent.DispatchHooks {
+	return coreevent.DispatchHooks{
+		Spawn: func(start bool, owner any, call func()) {
+			gco.CreateAndStart(start, owner, func(coroutine.Thread) int {
+				call()
+				return 0
+			})
+		},
+		Wait: engine.WaitToDo,
 	}
 }
 
@@ -239,7 +565,7 @@ func (p *Game) doWhenLeftButtonDown(ev *eventLeftButtonDown) {
 			target.doWhenClick(target)
 		},
 		DispatchStage: func() {
-			p.sinkMgr.doWhenClick(p)
+			p.scriptEvents.doWhenClick(p)
 		},
 	})
 }
